LayoutTests/platform/mac/accessibility/visible-character-range-basic-expected.txt

This tests that visibleCharacterRange returns expected visible ranges.

Range of text-less div: NSRange: {0, 0}
Range of text div with default view size: NSRange: {0, 4495}
Range with view 0x500, scrollTop 0: NSRange: {9223372036854775807, 0}
Range with view 500x0, scrollTop 0: NSRange: {9223372036854775807, 0}
Range with view 1x500, scrollTop 0: NSRange: {9223372036854775807, 0}
Range with view 500x1, scrollTop 0: NSRange: {9223372036854775807, 0}
Range with view 80x80, scrollTop 0: NSRange: {0, 10}
Range with view 500x200, scrollTop 0: NSRange: {0, 810}

PASS successfullyParsed is true

TEST COMPLETE
